Update main.css
im trying to use the css main to change the height of all the photos so i dont have to put them in manually. if one of you guys knows how to do it please do it. i thought the way i did it would work but it isnt working.